+++ include/libpldm/state_set.h
@@ -149,13 +149,6 @@ enum pldm_state_set_availability_values {
 	PLDM_STATE_SET_AVAILABILITY_REBOOTING = 8
 };
 
-/* @brief List of states for the Operational Fault status (ID 10).
- */
-enum pldm_state_set_operational_fault_status_values {
-	PLDM_STATE_SET_OPERATIONAL_FAULT_STATUS_NORMAL = 1,
-	PLDM_STATE_SET_OPERATIONAL_FAULT_STATUS_STRESSED = 2,
-};
-
 /* @brief List of states for the Operational Stress status (ID 9).
  */
 enum pldm_state_set_operational_stress_status_values {